move quest command into sub package and extract record/enum
